Increased the locked alter attempt number to 3600.
The intention is to work around long exclusive lock waits on temporary
index cleanup with <9.2 versions where concurrent reindexing is not
implemented yet. So it would perform locked_alter_count *
locked_alter_timeout attempts before falling back to the temporary
index cleanup. The use case is shown in the issue #8.